NYSETL.Engines.E1.Cache (nys_etl v0.1.0)

Wrapper for Cachex, providing a large number of checksum caches under a supervision tree. While Cachex provides non-blocking reads and writes for caches, transactions for a specific cache block other transactions via a single GenServer message queue.

This supervision tree improves the concurrency of transactions by providing a cache ring of N caches, using :erlang.phash2/2 to deterministically assign each checksum to a member of the ring.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

clear()

dump(path)

get(checksum)

get(cache, key)

See Cachex.get/2.

load(path)

put!(checksum, value)

put!(cache, checksum, value)

start_link(init_arg)

transaction(checksum, fun)