NYSETL.Engines.E4.Transfer (nys_etl v0.1.0)

Functions

find_or_create_transferred_index_case_and_lab_results( index_case, destination_case_data, destination_county_id )

Looks for an existing index case for the provided case_id and county_id, and returns it if found.

Otherwise, creates a new index case and lab results that mirror an index case and lab results that have been transferred in commcare.

The data field of the new index case is the result of merging the data field from the old index case plus the data from the newly-transferred case in commcare.

The data fields of the lab results belonging to the new index case ONLY get data from the old lab results, not from the newly-transferred lab results in commcare, because the current thinking is that ECLRS data is more correct for lab results.