NYSETL.Extra.Map (nys_etl v0.1.0)

Extra shared functions for interacting with Map data structures.

Functions

merge_empty_fields(a, b)

Specs

merge_empty_fields(map(), map()) :: map()

Merge map b into map a, where only values in map a that are not nil or empty take precedence over values from map b.

Examples

iex> import NYSETL.Extra.Map
iex> %{a: 1} |> merge_empty_fields(%{a: 2})
%{a: 1}
iex> %{a: 1} |> merge_empty_fields(%{b: 2})
%{a: 1, b: 2}
iex> %{a: 1, b: nil} |> merge_empty_fields(%{b: 2})
%{a: 1, b: 2}
iex> %{a: 1, b: ""} |> merge_empty_fields(%{b: 2})
%{a: 1, b: 2}