NYSETL.Extra.Scrubber (nys_etl v0.1.0)

Functions to scrub all PII in ECLRS dump files so that it can be used in development. Sensitive data is replaced with random strings of the same length and format (letters for letters, digits for digits and dates for dates). Currently only works with the 45-column ECLRS files but can easily be extended.

To use, run

alias NYSETL.Extra.Scrubber
Scrubber.scrub_file("../eclrs/data/c201012_0932.txt", "./scrubbed.txt")

Some internal functions have been left public for your pleasure.

Functions

random_date(time)

scrub_and_remember(string, ets)

Return a string which is based on the input string but with all letters replaced with random uppercase letters and all digits replaced with random digits. Uses ETS to always return the same output for a certain input.

Examples

iex> import NYSETL.Extra.Scrubber
iex> ets = :ets.new(:scrubber, [:public])
iex> :rand.seed(:exsss, {101, 102, 103})
iex> scrub_and_remember("john doe", ets)
"XTKK RDF"
iex> scrub_and_remember("mary-ann doe", ets)
"EWZB-YLL ZCH"
iex> scrub_and_remember("john doe", ets)
"XTKK RDF"

scrub_file(input_path, output_path, ets \\ :ets.new(:scrubber, [:public]))

Make a copy of the input file with all sensitive data replaced with random data of the same type and length.

Uses ETS to make sure that a given source string is consistently replaced with the same scrubbed string. The caller can supply an existing ETS for testing purposes or similar.

Scrubber.scrub_file("../eclrs/data/c201012_0932.txt", "./scrubbed.txt")

scrub_if_index_should_be_scrubbed(atom, arg, ets)

Scrubs a string if its field index requires scrubbing the specified column layout.

Examples

iex> import NYSETL.Extra.Scrubber
iex> ets = :ets.new(:scrubber, [:public])
iex> :rand.seed(:exsss, {101, 102, 103})
iex> scrub_if_index_should_be_scrubbed(:v1, {"Doe", 0}, ets)
"XTK"
iex> scrub_if_index_should_be_scrubbed(:v1, {"59483", 8}, ets) # ZIP is not PII
"59483"

scrub_random(string)

Return a string which is based on the input string but with all letters replaced with random uppercase letters and all digits replaced with random digits. Replaces valid ISO dates with valid ISO dates.

Examples

iex> import NYSETL.Extra.Scrubber
iex> :rand.seed(:exsss, {101, 102, 103})
iex> scrub_random("john doe")
"XTKK RDF"
iex> scrub_random("john doe") # no memory
"EWZB YLL"
iex> scrub_random("123 West 27th St.")
"432 ZCHT 39OA QS."
iex> scrub_random("23MAY1965") # special handling of dates
"17FEB1945"
iex> scrub_random("23MAY1965:00:00:00.000000") # special handling of datetimes
"29NOV1994:00:00:00.000000"

scrub_row(arg, ets)

Scrub all PII fields in one row.

Examples

iex> import NYSETL.Extra.Scrubber
iex> ets = :ets.new(:scrubber, [:public])
iex> :rand.seed(:exsss, {101, 102, 103})
iex> scrub_row({:v1, "Doe|J|John|18MAR1965:00:00:00|||||||||||||||||||||||||||||||||||||||||unclear\n"}, ets)
"XTK|K|RDFE|28AUG2018:00:00:00|||||||||||||||||||||||||||||||||||||||||unclear\n"
iex> scrub_row({:v1, "Doe|A|Mary|23MAY1965:00:00:00.000000|||||||||||||||||||||||||||||||||||||||||POS"}, ets)
"XTK|Z|BYLL|18APR2016:00:00:00.000000|||||||||||||||||||||||||||||||||||||||||POS"