NYSETLWeb.Router.Helpers (nys_etl v0.1.0)

Module with named helpers generated from NYSETLWeb.Router.

Functions

health_check_path(conn_or_endpoint, action)

health_check_path(conn_or_endpoint, action, params)

health_check_url(conn_or_endpoint, action)

health_check_url(conn_or_endpoint, action, params)

live_dashboard_path(conn_or_endpoint, action)

live_dashboard_path(conn_or_endpoint, action, params)

live_dashboard_path(conn_or_endpoint, action, page, params)

live_dashboard_path(conn_or_endpoint, action, node, page, params)

live_dashboard_url(conn_or_endpoint, action)

live_dashboard_url(conn_or_endpoint, action, params)

live_dashboard_url(conn_or_endpoint, action, page, params)

live_dashboard_url(conn_or_endpoint, action, node, page, params)

oban_dashboard_path(conn_or_endpoint, action)

oban_dashboard_path(conn_or_endpoint, action, params)

oban_dashboard_path(conn_or_endpoint, action, page, params)

oban_dashboard_path(conn_or_endpoint, action, page, id, params)

oban_dashboard_url(conn_or_endpoint, action)

oban_dashboard_url(conn_or_endpoint, action, params)

oban_dashboard_url(conn_or_endpoint, action, page, params)

oban_dashboard_url(conn_or_endpoint, action, page, id, params)

page_path(conn_or_endpoint, action)

page_path(conn_or_endpoint, action, params)

page_url(conn_or_endpoint, action)

page_url(conn_or_endpoint, action, params)

path(data, path)

Generates the path information including any necessary prefix.

static_integrity(endpoint, path)

Generates an integrity hash to a static asset given its file path.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.